+++ @utahdts/utah-design-system-header/src/js/renderables/logoTitle/LogoTitle.js
@@ -14,10 +14,14 @@ import { getUtahHeaderSettings } from '../../settings/getUtahHeaderSettings';
  * @returns {Element}
  */
 export function LogoTitle() {
+  const utahHeaderSettings = getUtahHeaderSettings();
   const logoTitleURL = (
-    getUtahHeaderSettings().titleUrl
-    // @ts-ignore : backwards-compatibility - once upon a time, titleURL had been incorrectly cased...
-    || getUtahHeaderSettings().titleURL
+    // type says it can't be null, but for backwards-compatibility, it may be null
+    utahHeaderSettings.titleUrlAction?.actionUrl
+    // @ts-expect-error backwards-compatibility: used to could only be a string
+    || /** @type {string | undefined} */ (utahHeaderSettings.titleUrl)
+    // @ts-expect-error backwards-compatibility: once upon a time, titleURL had been incorrectly cased...
+    || /** @type {string | undefined} */ (utahHeaderSettings.titleURL)
   );
   const logoTitleWrapper = !logoTitleURL ? renderDOMSingle(LogoTitleWrapper) : renderDOMSingle(LogoTitleWrapperLink);
   if (!logoTitleWrapper) {
@@ -26,6 +30,13 @@ export function LogoTitle() {
   if (logoTitleURL) {
     logoTitleWrapper.setAttribute('href', logoTitleURL);
   }
+  // type says it can't be null, but for backwards-compatibility, it may be null
+  if (utahHeaderSettings.titleUrlAction?.actionFunction) {
+    if (logoTitleWrapper.onclick) {
+      throw new Error('LogoTitle: logoTitleWrapper already has an onclick');
+    }
+    logoTitleWrapper.onclick = utahHeaderSettings.titleUrlAction.actionFunction;
+  }
 
   // Render Logo image
   const logoWrapper = logoTitleWrapper.querySelector(getCssClassSelector(domConstants.TITLE__LOGO));
@@ -33,9 +44,9 @@ export function LogoTitle() {
     throw new Error('LogoTitle: logoWrapper is null');
   }
 
-  const settingsLogo = getUtahHeaderSettings().logo;
-  const settingsShowTitle = getUtahHeaderSettings().showTitle;
-  const settingsTitle = getUtahHeaderSettings().title;
+  const settingsLogo = utahHeaderSettings.logo;
+  const settingsShowTitle = utahHeaderSettings.showTitle;
+  const settingsTitle = utahHeaderSettings.title;
   if (settingsLogo) {
     /** @type {HTMLCollection | Element} */
     let settingsLogoElement;