+++ @utahdts/utah-design-system-header/src/js/settings/defaultSettings.js
@@ -24,6 +24,8 @@ export const defaultSettings = {
   size: sizes.MEDIUM,
   skipLinkUrl: '#main-content',
   title: 'My Utah.gov Site',
-  titleUrl: '/',
+  titleUrlAction: {
+    actionUrl: '/',
+  },
   utahId: false,
 };