+++ utah-design-system-website/src/main.jsx
@@ -1,13 +1,12 @@
-import { UtahDesignSystemContextProvider, UtahHeaderContextProvider } from '@utahdts/utah-design-system';
-import { defaultSettings } from '@utahdts/utah-design-system-header';
+import { UtahDesignSystemContextProvider } from '@utahdts/utah-design-system';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { App } from './App';
 import { AppContextProvider } from './react/context/AppContext/AppContextProvider';
 import { CssContextProvider } from './react/context/cssContext/CssContextProvider';
+import { WebsiteUtahHeaderContextProvider } from './react/context/WebsiteUtahHeaderContextProvider';
 import { notNull } from './react/util/notNull/notNull';
-import { websiteUtahHeaderSettings } from './websiteUtahHeaderSettings';
 
 const baseName = '';
 
@@ -17,9 +16,9 @@ ReactDOM.createRoot(notNull(document.getElementById('root'), 'main.jsx: root not
       <UtahDesignSystemContextProvider defaultSettings={{ bannerDuration: 3500 }}>
         <AppContextProvider>
           <CssContextProvider>
-            <UtahHeaderContextProvider defaultSettings={{ ...defaultSettings, ...websiteUtahHeaderSettings }}>
+            <WebsiteUtahHeaderContextProvider>
               <App />
-            </UtahHeaderContextProvider>
+            </WebsiteUtahHeaderContextProvider>
           </CssContextProvider>
         </AppContextProvider>
       </UtahDesignSystemContextProvider>