+++ utah-design-system-website/src/react/components/websiteContent/library/patterns/UtahHeaderDocumentation/UtahHeaderDocumentation.jsx
@@ -580,12 +580,31 @@ export function UtahHeaderDocumentation() {
 
             <TableRow>
               <TableCell>
-                <span className="prop__name"><a href="#section-config-titleUrl">titleUrl</a></span><br />
-                <span className="prop__types">string</span><br />
-                Default: <code>/</code>
+                <span className="prop__name"><a href="#section-config-titleUrlAction">titleUrlAction</a></span><br />
+                <span className="prop__types">TitleUrlAction</span><br />
               </TableCell>
               <TableCell>
-                When the logo and/or title are clicked, the browser will navigate to this URL.
+                Use to provide a url, and/or a function to apply, when the title is clicked.
+              </TableCell>
+            </TableRow>
+
+            <TableRow>
+              <TableCell>
+                <span className="prop__name"><a href="#section-config-titleUrlAction">titleUrlAction.titleUrl</a></span><br />
+                <span className="prop__types">string</span><br />
+              </TableCell>
+              <TableCell>
+                The url to show when the title is hovered. Will navigate to this url if <code>titleUrlAction</code> is not supplied.
+              </TableCell>
+            </TableRow>
+
+            <TableRow>
+              <TableCell>
+                <span className="prop__name"><a href="#section-config-titleUrlAction">titleUrlAction.titleUrlAction</a></span><br />
+                <span className="prop__types">function</span><br />
+              </TableCell>
+              <TableCell>
+                This function fires when the title is clicked. It is passed the event, allowing things like <code>preventDefault</code> and <code>stopPropagation</code>.
               </TableCell>
             </TableRow>
           </TableBody>
@@ -825,18 +844,30 @@ export function UtahHeaderDocumentation() {
         />
       </div>
 
-      <h4 id="section-config-titleUrl">titleUrl</h4>
+      <h4 id="section-config-titleUrl">titleUrlAction</h4>
       <div>
         <PreCodeForCodeString
           className="gray-block mt-spacing"
           codeRaw={`
+            To set the url for the title.
             setUtahHeaderSettings(
               {
                 ...other settings...,
-                titleUrl: '/',
+                titleUrlAction: { titleUrl: '/' },
               }
             )
-          `}
+
+            Or for a custom onclick action:
+            setUtahHeaderSettings(
+              {
+                ...other settings...,
+                titleUrlAction: { 
+                  titleUrl: '/', 
+                  actionFunction: () => { ... do something ... },
+                },
+              }
+            )
+            `}
         />
       </div>