+++ utah-design-system-website/test/react/components/websiteContent/library/patterns/UtahHeaderDocumentation/stringifyHeaderSettings.test.js
@@ -8,7 +8,8 @@ const MATCH_BASIC_FIELDS = [
   '"showTitle": true',
   '"size": "MEDIUM"',
   '"title": "Utah Design System"',
-  '"titleUrl": "/"',
+  '"titleUrlAction": {',
+  '"actionUrl": "/"',
   '"mediaSizes": {',
   '"mobile": 640,',
   '"tabletPortrait": 768,',
@@ -17,7 +18,7 @@ const MATCH_BASIC_FIELDS = [
 
 /**
  * @param {string[]} strings the strings to test to satisfy validity
- * @returns {function(string): boolean} true if all the strings match teh testStr
+ * @returns {function(string): boolean} true if all the strings match the testStr
  */
 function doesMatchStrings(strings) {
   return (resultStr) => strings.every((testStr) => resultStr.includes(testStr));
@@ -30,7 +31,7 @@ describe('stringifyHeaderSettings', () => {
       showTitle: true,
       size: sizes.MEDIUM,
       title: 'Utah Design System',
-      titleUrl: '/',
+      titleUrlAction: { actionUrl: '/' },
       mainMenu: {
         menuItems: [],
         title: 'test-menu',
@@ -52,7 +53,7 @@ describe('stringifyHeaderSettings', () => {
       showTitle: true,
       size: sizes.MEDIUM,
       title: 'Utah Design System',
-      titleUrl: '/',
+      titleUrlAction: { actionUrl: '/' },
       mainMenu: {
         menuItems: [],
         title: 'test-menu',
@@ -90,7 +91,7 @@ describe('stringifyHeaderSettings', () => {
       showTitle: true,
       size: sizes.MEDIUM,
       title: 'Utah Design System',
-      titleUrl: '/',
+      titleUrlAction: { actionUrl: '/' },
       mainMenu: {
         menuItems: [],
         title: 'test-menu',
@@ -122,7 +123,7 @@ describe('stringifyHeaderSettings', () => {
       showTitle: true,
       size: sizes.MEDIUM,
       title: 'Utah Design System',
-      titleUrl: '/',
+      titleUrlAction: { actionUrl: '/' },
       mainMenu: {
         menuItems: [],
         title: 'test-menu',
@@ -153,7 +154,7 @@ describe('stringifyHeaderSettings', () => {
       showTitle: true,
       size: sizes.MEDIUM,
       title: 'Utah Design System',
-      titleUrl: '/',
+      titleUrlAction: { actionUrl: '/' },
       mainMenu: {
         menuItems: [],
         title: 'test-menu',
@@ -197,7 +198,7 @@ describe('stringifyHeaderSettings', () => {
       showTitle: true,
       size: sizes.MEDIUM,
       title: 'Utah Design System',
-      titleUrl: '/',
+      titleUrlAction: { actionUrl: '/' },
       mainMenu: {
         menuItems: [],
         title: 'test-menu',